backend/requirements.txt
fastapi==0.110.1
uvicorn==0.30.1
pydantic==1.10.15
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
pymongo==4.7.2
dnspython==2.6.1
openai==1.35.0
faiss-cpu==1.8.0
sentence-transformers==2.7.0

